MR-94: Add the trainers form
